pyproject.toml
[project]
name = "financial-tracker"
version = "0.0.1"
description = "Financial tracking tool for bookkeeping, planning, projection, etc."
keywords = [
  "personal-finance",
  "bookkeeping",
  "budgeting",
  "financial-planning",
  "financial-projection",
  "artificial-intelligence",
  "large-language-model",
  "llm",
  "openai",
  "langchain",
]
readme = {file = "README.md", content-type = "text/markdown"}
license = {file = "LICENSE"}
authors = [
  {name = "Kevin Lin", email = "klinime@gmail.com"},
]
requires-python = ">=3.12"
dependencies = [
  "six>=1.16.0",
  "pandas",
  "PyMuPDF",
  "pdfservices-sdk",
  "langchain-openai",
  "langchain-community",
  "python-dotenv",
]

[project.optional-dependencies]

[tool.isort]
profile = "black"

[tool.black]
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
strict = true
disallow_untyped_defs = true
ignore_missing_imports = true
show_error_codes = true
